refactor : 요구사항에 따라 Lotto클래스 제자리(domain)로 이동
이유
: 요구사항(로또클래스는 이동해도 된다)
